+++ src/ChaiLove.cpp
@@ -58,6 +58,7 @@ bool ChaiLove::load(const std::string& file) {
 		sound.load();
 	}
 
+	console.load();
 	graphics.load();
 	image.load();
 	keyboard.load();
@@ -131,6 +132,9 @@ void ChaiLove::draw() {
 		#ifdef __HAVE_TESTS__
 		test.draw();
 		#endif
+
+		// Render the in-game console.
+		console.draw();
 
 		// Flip the buffer.
 		if (SDL_Flip(screen) == -1) {